+++ CHANGELOG.md
@@ -38,6 +38,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Allow filtering by entity names in the World Inspector (#1106, **@diogomsmiranda**).
 - Possibility of building the core library as a shared library (#1052, **@RiscadoA**).
 - Possibility of building the engine library as a shared library (#1053, **@RiscadoA**).
+- Plugins system argument, which can be used to dynamically add and remove plugins (#1034, **@RiscadoA**).
 
 ### Changed